+++ app/converters/gbfs_https_to_http.py
@@ -18,6 +18,7 @@ class GbfsHttpsToHttpConverter(BaseConverter):
         'mds.bird.co',
         'gbfs.prod.sharedmobility.ch',
         'api.voiapp.io',
+        'gbfs.api.ridedott.com',
     ]
 
     def convert(self, data: Union[dict, list], path: str) -> Union[dict, list]: